ws54: Modal component refactor.
