+++ frontend/src/models/PostData.ts
@@ -1,0 +1,9 @@
+interface PostData {
+  imageUrl: string;
+  species: string;
+  username: string;
+  sightingDate: string;
+  likes: number;
+}
+
+export default PostData;
